+++ jest.setup.ts
@@ -1,0 +1,9 @@
+import "@testing-library/jest-native/extend-expect";
+
+jest.mock("react-native", () => {
+  const actual = jest.requireActual("react-native");
+  return {
+    ...actual,
+    useColorScheme: () => "light",
+  };
+});